fix(v2.1): correct RV64 PC and address handling (#2920)
## Summary

- Fix RV64 AUIPC sign-extension so `pc + sign_extend(imm << 12)` is
handled consistently in execution, AIR, CUDA tracegen, and RVR
lifting/emission.
- Fix JALR and load/store effective-address handling so RV64 register
values are not silently truncated or wrapped into the implemented
PC/memory address space.
- Restore missing load/store carry/sign constraints and add negative
coverage for effective-address overflow.
- Harden RVR dynamic jumps so computed RV64 targets are bounds-checked
before narrowing to the dispatch-table PC type.

resolves int-8294

+++ crates/rvr/rvr-openvm-ir/src/instr.rs
@@ -101,10 +101,10 @@ pub enum Instr {
         rd: Reg,
         value: u32,
     },
-    /// Add upper immediate to PC (rd = pc + (imm << 12), pre-computed).
+    /// Add upper immediate to PC (rd = pc + sign_extend(imm << 12), pre-computed).
     Auipc {
         rd: Reg,
-        value: u32,
+        value: u64,
     },
     /// Load from memory.
     Load {
